moved Seo into own folder
