+++ frontend/src/pages/apps.js
@@ -3,7 +3,7 @@ import { Router } from "@reach/router"
 import "bootstrap/dist/css/bootstrap.min.css"
 
 import Layout from "../components/Layout/Layout"
-import SEO from "../components/Seo"
+import SEO from "../components/Seo/Seo"
 import AppGrid from "../components/AppGrid/AppGrid"
 import GamesPage from "./apps/games"
 import ShoppingPage from "./apps/shopping"
