+++ frontend/src/pages/apps/productivity.js
@@ -1,5 +1,5 @@
 import React from "react"
-import SEO from "../../components/Seo"
+import SEO from "../../components/Seo/Seo"
 
 const ProductivityPage = () => (
   <React.Fragment>
