+++ frontend/src/pages/favorite-apps.js
@@ -2,7 +2,7 @@ import React from "react"
 import "bootstrap/dist/css/bootstrap.min.css"
 
 import Layout from "../components/Layout/Layout"
-import SEO from "../components/Seo"
+import SEO from "../components/Seo/Seo"
 
 const FavoriteAppsPage = () => (
   <Layout>
