+++ frontend/src/pages/index.js
@@ -1,7 +1,7 @@
 import React from "react"
 
 import Layout from "../components/Layout/Layout"
-import SEO from "../components/Seo"
+import SEO from "../components/Seo/Seo"
 import "bootstrap/dist/css/bootstrap.min.css"
 
 import AppGrid from "../components/AppGrid/AppGrid"
