+++ frontend/src/pages/login.js
@@ -1,6 +1,6 @@
 import React from "react"
 import Layout from "../components/Layout/Layout"
-import SEO from "../components/Seo"
+import SEO from "../components/Seo/Seo"
 
 import UserContext from "../shared/UserContext"
 import * as Url from "../shared/urls"
