+++ frontend/src/pages/settings.js
@@ -1,7 +1,7 @@
 import React from "react"
 
 import Layout from "../components/Layout/Layout"
-import SEO from "../components/Seo"
+import SEO from "../components/Seo/Seo"
 
 const SettingsPage = () => (
   <Layout>
